Moved B2 flamer to main B2 blaster addon
Updated inheritance for B2Blaster_flamer, uses base B2 blaster.
Burn 'Em compat
