+++ addons/compats/weapons_burnEm/CfgMagazines.hpp
@@ -3,4 +3,8 @@ class CfgMagazines {
     class CLASS(Mag_150Rnd_X42): CLASS(Mag_Base) {
         ammo = "Flamethrower_Fuel";
     };
+
+    class CLASS(Mag_150rnd_B2Blaster_flamer): CLASS(Mag_Base) {
+        ammo = "Flamethrower_Fuel";
+    };
 };
