+++ addons/weapons/b2blaster/CfgWeapons.hpp
@@ -205,4 +205,22 @@ class CfgWeapons {
 
         muzzles[] = {"RocketLauncher"};
     };
+
+    class CLASS(B2Blaster_flamer): CLASS(B2Blaster_base) {
+        SCOPE_PUBLIC;
+
+        displayName = "[KC] Wrist Mounted Flamethrower";
+        descriptionShort = "Built in flamethrower";
+        baseWeapon = QCLASS(B2Blaster_flamer);
+
+        magazines[] = {QCLASS(Mag_150rnd_B2Blaster_flamer)};
+
+        canShootInWater = FALSE;
+
+        WBK_BurnEm_IsFlamethrower = "True";
+        WBK_BurnEm_FlamethrowerDistance = 40;
+        WBK_BurnEm_RequiredBackpack = "";
+        WBK_BurnEm_flamethrowerParticlePos = "[[0, 0, 0], 'righthand']";
+        WBK_BurnEm_FlamethrowerSoundArray = "['flamethrower_start', 'flamethrower_looping', 4.8, 'flamethrower_end']";
+    };
 };
