--- addons/weapons/b2blaster_flamer/script_component.hpp
@@ -1,4 +1,0 @@
-#define SUBCOMPONENT b2blaster_flamer
-#define SUBCOMPONENT_BEAUTIFIED B2 Blaster (Flamer)
-
-#include "..\script_component.hpp"
